producer_api/requirements.txt
Flask>=1.1.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
kafka-python>=1.4.7
requests>=2.22.0
pytest>=5.1.2
